@@ -1,9 +1,9 @@
 {
   "name": "angular-starter",
-  "version": "9.0.5",
+  "version": "9.1.2",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --port 4200",
+    "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -11,6 +11,7 @@
     "dev:ssr": "ng run angular-starter:serve-ssr",
     "serve:ssr": "node dist/angular-starter/server/main.js",
     "build:ssr": "ng build --prod && ng run angular-starter:server:production",
+    "build-hub:ssr": "ng build --prod --base-href=/angular-example-seo/ && ng run angular-starter:server:production    ",
     "prerender": "ng run angular-starter:prerender"
   },
   "repository": {
@@ -29,48 +30,46 @@
   "homepage": "https://github.com/ganatan/angular-example-seo#readme",
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.0.5",
-    "@angular/common": "9.0.5",
-    "@angular/compiler": "9.0.5",
-    "@angular/core": "9.0.5",
-    "@angular/forms": "9.0.5",
-    "@angular/platform-browser": "9.0.5",
-    "@angular/platform-browser-dynamic": "9.0.5",
-    "@angular/platform-server": "9.0.5",
-    "@angular/router": "9.0.5",
-    "@angular/service-worker": "9.0.5",
-    "@fortawesome/fontawesome-free": "5.12.1",
-    "@nguniversal/express-engine": "9.0.1",
+    "@angular/animations": "9.1.2",
+    "@angular/common": "9.1.2",
+    "@angular/compiler": "9.1.2",
+    "@angular/core": "9.1.2",
+    "@angular/forms": "9.1.2",
+    "@angular/platform-browser": "9.1.2",
+    "@angular/platform-browser-dynamic": "9.1.2",
+    "@angular/platform-server": "9.1.2",
+    "@angular/router": "9.1.2",
+    "@angular/service-worker": "9.1.2",
+    "@fortawesome/fontawesome-free": "5.13.0",
+    "@nguniversal/express-engine": "9.1.0",
     "bootstrap": "4.4.1",
-    "chart.js": "2.9.3",
     "express": "4.17.1",
     "jquery": "3.4.1",
-    "ng2-charts": "2.3.0",
-    "rxjs": "6.5.4",
+    "rxjs": "6.5.5",
     "tslib": "1.11.1",
-    "zone.js": "0.10.2"
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.900.5",
-    "@angular/cli": "9.0.5",
-    "@angular/compiler-cli": "9.0.5",
-    "@angular/language-service": "9.0.5",
-    "@nguniversal/builders": "9.0.1",
-    "@types/express": "4.17.3",
-    "@types/node": "13.9.0",
-    "@types/jasmine": "3.5.8",
+    "@angular-devkit/build-angular": "0.901.1",
+    "@angular/cli": "9.1.1",
+    "@angular/compiler-cli": "9.1.2",
+    "@angular/language-service": "9.1.2",
+    "@nguniversal/builders": "9.1.0",
+    "@types/express": "4.17.6",
+    "@types/node": "13.13.0",
+    "@types/jasmine": "3.5.10",
     "@types/jasminewd2": "2.0.8",
-    "codelyzer": "5.2.1",
+    "codelyzer": "5.2.2",
     "jasmine-core": "3.5.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.4.1",
+    "jasmine-spec-reporter": "5.0.1",
+    "karma": "5.0.2",
     "karma-chrome-launcher": "3.1.0",
     "karma-coverage-istanbul-reporter": "2.1.1",
     "karma-jasmine": "3.1.1",
-    "karma-jasmine-html-reporter": "1.5.2",
-    "protractor": "5.4.3",
-    "ts-node": "8.6.2",
-    "tslint": "6.0.0",
-    "typescript": "3.7.5"
+    "karma-jasmine-html-reporter": "1.5.3",
+    "protractor": "5.4.4",
+    "ts-node": "8.8.2",
+    "tslint": "6.1.1",
+    "typescript": "3.8.3"
   }
 }
