@@ -3,14 +3,14 @@ import { CommonModule } from '@angular/common';
 
 import { ItemsComponent } from './items.component';
 import { ItemsRoutingModule } from './items-routing.module';
-// import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
 import { ItemsService } from './items.service';
 
 @NgModule({
   imports: [
     CommonModule,
     ItemsRoutingModule,
-//    HttpClientModule
+    HttpClientModule
   ],
   exports: [
     ItemsComponent
