@@ -4,7 +4,31 @@ import { Routes, RouterModule } from '@angular/router';
 import { ContactComponent } from './contact.component';
 
 const routes: Routes = [
-  { path: '', component: ContactComponent },
+  {
+    path: '', component: ContactComponent, children: [
+      {
+        path: 'mailing',
+        loadChildren: () => import(`./mailing/mailing.module`)
+          .then(m => m.MailingModule)
+      },
+      {
+        path: '',
+        loadChildren: () => import(`./mailing/mailing.module`)
+          .then(m => m.MailingModule)
+      },
+      {
+        path: 'map',
+        loadChildren: () => import(`./map/map.module`)
+          .then(m => m.MapModule)
+      },
+      {
+        path: '**',
+        loadChildren: () => import(`./mailing/mailing.module`)
+          .then(m => m.MailingModule)
+      },
+
+    ]
+  },
 ];
 
 @NgModule({
