@@ -22,70 +22,40 @@ export class HomeComponent implements OnInit {
     this.features =
       [
         {
-          name: 'Template Driven Forms',
-          description: 'Build an Angular form with a component and template',
-          icon: 'far fa-file-alt',
-          link: 'template-driven-forms'
+          name: 'Bootstrap components',
+          description: 'Bootstrap components Description',
+          icon: 'fab fa-bootstrap',
+          link: 'bootstrap-components'
         },
         {
-          name: 'Chart.js',
-          description: 'Simple yet flexible JavaScript charting for designers & developers',
-          icon: 'fa-share-alt-square',
-          link: 'chartjs'
-        },
-        {
-          name: 'Components',
-          description: 'Define and control views',
-          icon: 'fa-share-alt-square',
-          link: 'components'
+          name: 'Reactive Form',
+          description: 'Reactive Form Description',
+          icon: 'fab fa-bootstrap',
+          link: 'reactive-form'
         },
         {
           name: 'Services',
-          description: 'A great way to share information among classes',
-          icon: 'fa-address-card',
+          description: 'Services Description',
+          icon: 'fab fa-bootstrap',
           link: 'services'
         },
         {
-          name: 'HttpClient',
-          description: 'HttpClient',
-          icon: 'fa-comment-alt',
+          name: 'Components',
+          description: 'Components Description',
+          icon: 'fab fa-bootstrap',
+          link: 'components'
+        },
+        {
+          name: 'Template Driven Forms',
+          description: 'Template Driven Forms Description',
+          icon: 'fab fa-bootstrap',
+          link: 'template-driven-forms'
+        },
+        {
+          name: 'Items',
+          description: 'Items',
+          icon: 'fab fa-bootstrap',
           link: 'httpclient'
-        },
-        {
-          name: 'Directives',
-          description: 'Change the appearance or behavior of a DOM element',
-          icon: 'fa-text-width',
-          link: 'directives'
-        },
-        {
-          name: 'Pipes',
-          description: 'Write display-value transformations',
-          icon: 'fa-code',
-          link: 'pipes'
-        },
-        {
-          name: 'RxJS/Observables',
-          description: 'Provide support for passing messages between publishers and subscribers',
-          icon: 'fa-comment-alt',
-          link: 'observables'
-        },
-        {
-          name: 'modal',
-          description: 'Add dialogs to your site',
-          icon: 'fa-comment-alt',
-          link: 'modal'
-        },
-        {
-          name: 'Reactiveform',
-          description: 'Create Reactive Form',
-          icon: 'fa-comment-alt',
-          link: 'reactiveform'
-        },
-        {
-          name: 'Template Driven Form',
-          description: 'Create Template Driven Form',
-          icon: 'fa-comment-alt',
-          link: 'templatedriven'
         },
       ];
 
@@ -104,4 +74,5 @@ export class HomeComponent implements OnInit {
           ' Elle applique le Routing, le Lazy loading, le Server side rendering et les Progressive Web App (PWA)'
       });
   }
+
 }
