@@ -3,8 +3,8 @@ export const environment = {
   application:
   {
     name: 'angular-starter',
-    angular: 'Angular 9.0.5',
+    angular: 'Angular 9.1.2',
     bootstrap: 'Bootstrap 4.4.1',
-    fontawesome: 'Font Awesome 5.12.1',
+    fontawesome: 'Font Awesome 5.13.0',
   }
 };
